Implement DI and Link services
great momentum tonight. 
Adjusted the dependency injection container to be a global service, and added the linking service intot he mix

added a few interfaces to help make linking a bit easier. There is a little configuration and a little convention in the mix, but I think it will work well. 

Now the main question is to restrinct linking to its own api endpoint, as metadata when saving items, or both...